quicksave

+++ Joystick.py
@@ -73,7 +73,7 @@ class Joystick:
     def is_connected(self):
         return self.connected
     
-    def get_axis(self, axis_name: str, limit_perc: int = 100):
+    def get_axis(self, axis_name: str, limit_perc: float = 100):
         if axis_name in XBOX_AXES:
             return self.joystick.get_axis(XBOX_AXES[axis_name]) * (limit_perc / 100)
         else:
